refact(redux): rename redundant files names
